+++ src/panser.rs
@@ -17,7 +17,7 @@ use std::path::Path;
 use std::str::{self, FromStr};
 use std::sync::mpsc;
 use std::thread;
-use super::{Error, FromFormat, Result, ToFormat};
+use super::{Error, Framing, FromFormat, Result, ToFormat};
 
 /// Convert the input in a format to the output format.
 ///
@@ -27,7 +27,7 @@ use super::{Error, FromFormat, Result, ToFormat};
 /// or a method that deserializes from a reader, but all currently supported serialization
 /// libraries do support the `from_slice` or similar method for deserialization from a slice of
 /// bytes. 
-pub fn transcode<W: Write>(input: &[u8], output: &mut W, from: FromFormat, to: ToFormat, framed: bool, include_newline: bool) -> Result<()> {
+pub fn transcode<W: Write>(input: &[u8], output: &mut W, from: FromFormat, to: ToFormat, framing: Option<Framing>) -> Result<()> {
     let value = {
         match from {
             FromFormat::Bincode => bincode::deserialize::<serde_json::Value>(input)?,
@@ -59,54 +59,90 @@ pub fn transcode<W: Write>(input: &[u8], output: &mut W, from: FromFormat, to: T
             ToFormat::Yaml => serde_yaml::to_vec(&value)?,
         }
     };
-    if framed {
-        let mut frame_length = [0; 4];
-        BigEndian::write_u32(&mut frame_length, encoded_data.len() as u32);
-        output.write(&frame_length)?;
+    if let Some(f) = framing {
+        match f {
+            Framing::Sized => {
+                let mut frame_length = [0; 4];
+                BigEndian::write_u32(&mut frame_length, encoded_data.len() as u32);
+                output.write(&frame_length)?;
+            },
+            _ => {},
+        }
     }
     output.write(&encoded_data)?;
-    if include_newline {
-        output.write(&[b'\n'])?;
+    if let Some(f) = framing {
+        match f {
+            Framing::Delimited(delimiter) => {
+                output.write(&[delimiter; 1])?;
+            },
+            _ => {},
+        }
     }
     output.flush()?;
     Ok(())
 }
 
-fn read<R: BufRead>(mut reader: R, framed: bool, message_tx: mpsc::Sender<Vec<u8>>) -> Result<()> {
-    if framed {
-        // If the input is framed, this means messages (input data) can be continuous read from
-        // stdin or similar stream until End of File (EOF) is reached. Reaching EOF typically
-        // occurs in a streaming capacity when stdout is closed by the proceeding application that
-        // is piping (|) the output to the stdin of this application. Since the data is framed, the
-        // application knows the number of bytes to read for each message without having to look
-        // for a delimiting bytes, such as a newline (0x0A), or read the entire file to reach EOF
-        // before transcoding. In other words, framed messages can be read as they stream into the
-        // application.
-        loop {
-            let mut frame_length_buf = [0; 4];
-            reader.read_exact(&mut frame_length_buf).map_err(|e| {
-                match e.kind() {
-                    ErrorKind::UnexpectedEof => Error::Eof,
-                    _ => Error::Io(e)
-                }
-            })?;
-            let mut frame_length_cursor = Cursor::new(frame_length_buf);
-            let frame_length = frame_length_cursor.read_u32::<BigEndian>()?;
-            let mut buf = vec![0; frame_length as usize];
-            reader.read_exact(&mut buf).map_err(|e| {
-                match e.kind() {
-                    ErrorKind::UnexpectedEof => Error::Eof,
-                    _ => Error::Io(e)
-                }
-            })?;
-            message_tx.send(buf).unwrap();
+/// Reads exact length of bytes. 
+///
+/// This assumes the first four bytes of a message are the total data
+/// length encoded as an unsigned 32-bit integer in Big Endian (Network Order). Reading continues
+/// until the End-of-File (EOF) is reached.
+///
+/// Since the data is framed, the application can read messages as they as they are "streamed" into
+/// the reader without having to read the entire stream or file into memory. Messages can be
+/// transcoded as they arrive and continuous written to output.
+fn read_exact<R: BufRead>(mut reader: R, message_tx: mpsc::Sender<Vec<u8>>) -> Result<()> {
+    loop {
+        let mut frame_length_buf = [0; 4];
+        reader.read_exact(&mut frame_length_buf).map_err(|e| {
+            match e.kind() {
+                ErrorKind::UnexpectedEof => Error::Eof,
+                _ => Error::Io(e)
+            }
+        })?;
+        let mut frame_length_cursor = Cursor::new(frame_length_buf);
+        let frame_length = frame_length_cursor.read_u32::<BigEndian>()?;
+        let mut buf = vec![0; frame_length as usize];
+        reader.read_exact(&mut buf).map_err(|e| {
+            match e.kind() {
+                ErrorKind::UnexpectedEof => Error::Eof,
+                _ => Error::Io(e)
+            }
+        })?;
+        message_tx.send(buf).unwrap();
+    }
+}
+
+/// Reads until a delimiter is reached.
+///
+/// Reading continues until the End-of-File (EOF) is reached.
+///
+/// Since the data is framed, the application can read messages as they as they are "streamed" into
+/// the reader without having to read the entire stream or file into memory. Messages can be
+/// transcoded as they arrive and continuous written to output.
+fn read_until<R: BufRead>(mut reader: R, delimiter: u8, message_tx: mpsc::Sender<Vec<u8>>) -> Result<()> {
+    loop {
+        let mut buf = Vec::new();
+        reader.read_until(delimiter, &mut buf).map_err(|e| {
+            match e.kind() {
+                ErrorKind::UnexpectedEof => Error::Eof,
+                _ => Error::Io(e)
+            }
+        })?;
+        message_tx.send(buf).unwrap();
+    }
+}
+
+fn read<R: BufRead>(mut reader: R, framing: Option<Framing>, message_tx: mpsc::Sender<Vec<u8>>) -> Result<()> {
+    if let Some(f) = framing {
+        match f {
+            Framing::Sized => read_exact(reader, message_tx)?,
+            Framing::Delimited(delimiter) => read_until(reader, delimiter, message_tx)?,
         }
     } else {
-        // If the input is _not_ framed, the asumption is currently that the input is from a file,
-        // either from a redirection (<) or as a first optional ARG from the command line. In the
-        // case of a file, the entire contents should be read and then transcoded. This is not very
-        // efficient, but unless framing is used or a delimiter is used within the file to separate
-        // messages, this is the only universal way to transcode all serialization formats.
+        // If framing is not used, then the end stream or file must be read before transcoding
+        // begins. This is the only really universal way to transcode a non-framed stream since not
+        // all serialization formats use framing.
         let mut buf = Vec::new();
         let bytes_count = reader.read_to_end(&mut buf)?;
         if bytes_count > 0 {
@@ -115,10 +151,6 @@ fn read<R: BufRead>(mut reader: R, framed: bool, message_tx: mpsc::Sender<Vec<u8
             }
         } 
     }
-    // TODO: Add delimited-based reading. The `read_until` function should be used instead of
-    // framing or the `read_to_end` function. This would allow reading multiple messages from
-    // a file or stream (stdin) without having to read the entire contents before transcoding,
-    // similar to the frame reading.
     Ok(())
 }
 
@@ -132,12 +164,13 @@ fn read<R: BufRead>(mut reader: R, framed: bool, message_tx: mpsc::Sender<Vec<u8
 /// extension if a file is provided and it has an extension. The default input format is JSON. If
 /// the input format is not JSON and a file with an appropriate extension is _not_ used, then the
 /// `from` parameter should not be `None`. A similar procedure is used for the output (to) format.
-pub fn run(input: Option<&str>, output: Option<&str>, from: Option<FromFormat>, to: Option<ToFormat>, framed_input: bool, framed_output: bool, include_newline: bool) -> Result<()> {
+pub fn run(input: Option<&str>, output: Option<&str>, from: Option<FromFormat>, to: Option<ToFormat>, input_framing: Option<Framing>, output_framing: Option<Framing>) -> Result<()> {
     let (message_tx, message_rx) = mpsc::channel::<Vec<u8>>();
-    // Use `BufRead` instead of `Read` to add additionally reading methods, like `read_until`.
+    // Use `BufRead` instead of `Read` to add additional reading methods, like `read_until`. The
+    // `Send` trait is needed to move the reader to the read thread.
     let reader: Box<BufRead + Send> = {
-        if let Some(p) = input {
-            Box::new(BufReader::new(File::open(p)?))
+        if let Some(i) = input {
+            Box::new(BufReader::new(File::open(i)?))
         } else {
             Box::new(BufReader::new(io::stdin()))
         }
@@ -150,7 +183,7 @@ pub fn run(input: Option<&str>, output: Option<&str>, from: Option<FromFormat>, 
         }
     };
     let handle = thread::spawn(move || {
-        read(reader, framed_input, message_tx).or_else(|e| {
+        read(reader, input_framing, message_tx).or_else(|e| {
             match e {
                 Error::Eof => {
                     Ok(())
@@ -187,8 +220,7 @@ pub fn run(input: Option<&str>, output: Option<&str>, from: Option<FromFormat>, 
                     ToFormat::Msgpack
                 }
             }), 
-            framed_output,
-            include_newline)?;
+            output_framing)?;
         } else {
             break;
         }